% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{generatePermutations}
\alias{generatePermutations}
\title{Generate permutations}
\usage{
generatePermutations(n, K)
}
\arguments{
\item{n}{total number of elements to permute}

\item{K}{size of the subgroup}
}
\value{
matrix of permutation indices, each row represents a permutation
}
\description{
generate a subgroup of permutations of S_n of cardinality K
}
