% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{NaivePermutationTest}
\alias{NaivePermutationTest}
\title{Naive permutation test}
\usage{
NaivePermutationTest(X, Y, K = 999)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{K}{number of permutations}
}
\value{
p-value of the test
}
\description{
Implements the naive permutation test for testing whether the first
regression coefficient associated with Y~X is zero.
}
