% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{ResidualPermutationTest}
\alias{ResidualPermutationTest}
\title{Residual permutation test}
\usage{
ResidualPermutationTest(X, Y, K = 99)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{K}{number of permutations}
}
\value{
p-values for RPT and RPT_em
}
\description{
Implements the residual permutation test for testing whether the first
regression coefficient associated with Y~X is zero.
}
