% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{ResidualPermutationTest}
\alias{ResidualPermutationTest}
\title{Residual permutation test}
\usage{
ResidualPermutationTest(X, Y, K = 99)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{K}{number of permutations}
}
\value{
p-values for RPT and RPT_em
}
\description{
Implements the residual permutation test for testing whether the first
regression coefficient associated with Y~X is zero.
}
\examples{
vector.normalise <- function(v){v / sqrt(sum(v^2))}
n <- 600; p <- 100; k <- 5; coeff_sig <- 1; sigma <- 1; AR <- 0; 
coeff_shape='gaussian'
X <- matrix(rnorm(n*p), n)
alpha <- vector.normalise(c(rnorm(k), rep(0, p-k)))
beta <- vector.normalise(c(rnorm(k), rep(0, p-k)))
e <- rt(n, df=2)
eps <- rt(n, df=2)

b <- 0.5 # coefficient of interest
Z <- X \%*\% alpha + e
Y <- X \%*\% beta + b * Z + eps

X_aug <- cbind(Z, X)
NaivePermutationTest(X_aug, Y, K=99)
ResidualPermutationTest(X_aug, Y, K=99)
}
